requirements.txt
django~=4.2.6
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.0

